+++ .gitignore
@@ -3,3 +3,5 @@ dist/
 site/data/webpack.json
 /npm-debug.log
 yarn-error.log
+.idea
+package-lock.json
