+++ site/layouts/partials/nav.html
@@ -7,13 +7,13 @@
 
 	<!-- Primary Nav -->
 	<ul class="flex b grey-3">
-		<li><a href="/products" class="pa3 no-underline db">Products</a></li>
+		<li><a href="/projekte" class="pa3 no-underline db">Projekte</a></li>
 
-		<li><a href="/values" class="pa3 no-underline db">Values</a></li>
+		<li><a href="/ziele" class="pa3 no-underline db">Ziele</a></li>
 
 		<li><a href="/post" class="pa3 no-underline db">Blog</a></li>
 
-		<li><a href="/contact" class="pa3 no-underline db">Contact</a></li>
+		<li><a href="/mitmachen" class="pa3 no-underline db">Mitmachen</a></li>
 	</ul>
 
 </nav>
