+++ src/js/cms.js
@@ -6,14 +6,14 @@ import styles from "!to-string-loader!css-loader!postcss-loader!sass-loader!../c
 
 import HomePreview from "./cms-preview-templates/home";
 import PostPreview from "./cms-preview-templates/post";
-import ProductsPreview from "./cms-preview-templates/products";
-import ValuesPreview from "./cms-preview-templates/values";
-import ContactPreview from "./cms-preview-templates/contact";
+import ProjektePreview from "./cms-preview-templates/projekte";
+import ValuesPreview from "./cms-preview-templates/ziele";
+import ContactPreview from "./cms-preview-templates/mitmachen";
 
 CMS.registerPreviewStyle(styles, { raw: true });
 CMS.registerPreviewTemplate("home", HomePreview);
 CMS.registerPreviewTemplate("post", PostPreview);
-CMS.registerPreviewTemplate("products", ProductsPreview);
-CMS.registerPreviewTemplate("values", ValuesPreview);
+CMS.registerPreviewTemplate("projekte", ProjektePreview);
+CMS.registerPreviewTemplate("ziele", ValuesPreview);
 CMS.registerPreviewTemplate("contact", ContactPreview);
 CMS.init();
